// src/type.ts
export interface Property {
  name: string;
  value: any;
}

export interface SharedEventHandler {
  addEvent(root: EventTarget, el: EventTarget, type: string, callback: (e: Event) => void): void;
  removeEvent(el: EventTarget, type: string): void;
  replaceEvent(el: EventTarget, type: string, callback: (e: Event) => void): void;
  removeEvents(el: EventTarget): void;
}


export interface Stem {
  enterUnsafeUpdateZone(cb: () => void): void;
  setEventHandler(eventHandler: SharedEventHandler): void;
  getEventHandler(): SharedEventHandler;
  render(el: FuelElement, callback?: (el: FuelDOMNode) => void, context?: any, updateOwner?: boolean): void;
  registerOwner(el: FuelElement): void;
  owner(): FuelElement;
  unmountComponent(fuelElement: FuelElement, cb: () => void): void;
}
// ...
export interface FuelDOMNode extends EventTarget {
  removeAttribute(key: string): void;
  appendChild(child: FuelDOMNode): FuelDOMNode;
  parentNode: FuelDOMNode;
  removeChild(node: FuelDOMNode): void;
  replaceChild(newNode: FuelDOMNode, oldNode: FuelDOMNode): void
  childNodes: FuelDOMNode[];
  nodeType: number;
  children: FuelDOMNode[];
  style?: {[key: string]: string};
  textContent: string;
}


export type FuelComponentType = number|string|FuelComponentStatic<any, any>|StatelessComponent<any>;


export interface ESSubscription {
  unsubscribe(): void
}


export interface ESObservable<T> {
  subscribe(value: T): ESSubscription;
}


export interface PublicFuelElement {
  type: any;
  props: Property[],
  key: string|number;
  children: FuelElement[];
}


export interface FuelElement extends PublicFuelElement {
  dom: FuelDOMNode;
  _unmounted: boolean;
  _ownerElement: FuelElement,
  _stem?: Stem;
  _componentInstance?: FuelComponent<any, any>;
  _componentRenderedElementTreeCache?: FuelElement;
  _keymap?: {[key: string]: FuelElement};
  _subscriptions?: ESSubscription[];
}
// ...
export interface FuelComponent<Props, State> {
  state?: State;
  readonly props: Props;
  readonly context: any;
  readonly refs?: {[key:string]: FuelComponent<any, any>|Element};
  componentWillUnmount(): void;
  componentWillMount(): void;
  componentDidMount(): void;
  componentWillUpdate(): void;
  componentDidUpdate(): void;
  componentWillReceiveProps(props: Props): void;
  render(): JSX.Element;
  shouldComponentUpdate(nextProps: Props, prevProps: Props): boolean;
  setState(s: State, cb?: () => void): void;
  getChildContext(): any;
}
// ...
export const DOMEvents: {[key: string]: boolean} = eventmap`Copy Cut Paste CompositionEnd CompositionStart CompositionUpdate Focus Blur Change Input Submit Load Error KeyDown KeyPress KeyUp Abort CanPlay CanPlayThrough DurationChange Emptied Encrypted Ended LoadedData LoadedMetadata LoadStart Pause Play Playing Progress RateChange Seeked Seeking Stalled Suspend TimeUpdate VolumeChange Waiting Click ContextMenu DoubleClick Drag DragEnd DragEnter DragExit DragLeave DragOver DragStart Drop MouseDown MouseEnter MouseLeave MouseMove MouseOut MouseOver MouseUp Select TouchCancel TouchEnd TouchMove TouchStart Scroll Wheel`;
// ...
export interface Props<T> {
  children?: FuelElement;
  key?: string | number;
  ref?: string | ((component: T) => any);
}
// ...
    interface Element extends FuelElement { }